No autoplay on first question
Deactivates autoplay on sound for the first question, even if autoplay
is set, except when a title slide is set. This is to prevent audio from
playing when the page loads and allows the quiz maker to randomize
questions while having all but the first autoplay.

Also, added new version of first_second quiz that uses hangman-style
questions.